# public/openapi/project-board.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ProjectBoard API",
    "description": "API for managing Kanban boards, columns, cards, and tasks. Returns full board details including all columns, cards, members, and totals. The default_board_id in listBoards response indicates the most recently updated board - use this when user doesn't specify a board. Labels are global and can be used on any board.",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "https://your-domain.com",
      "description": "Server URL is set dynamically when accessed via /project-board/openapi.json"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "ProjectBoard API token (generate from Nova → ProjectBoard → API Tokens)"
      }
    },
...
  },
  "security": [
    { "bearerAuth": [] }
  ],
  "paths": {
...
    "/api/projects-board/cards/{id}/comments": {
      "post": {
        "operationId": "addComment",
        "summary": "Add a comment to a card",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["body"],
                "properties": {
                  "body": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Comment added" }
        }
      }
    },
    "/api/project-board/cards/{id}/attachments": {
      "post": {
        "operationId": "addAttachment",
        "summary": "Add an attachment to a card via URL",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["url"],
                "properties": {
                  "url": { "type": "string" },
                  "name": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Attachment added" }
        }
      }
    }
  }
}